mechanic_shop_api/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Marshmallow==1.4.0
Flask-Caching==2.3.0
Flask-Limiter==3.7.0
flask-swagger==0.2.14
flask-swagger-ui==5.21.0
marshmallow==3.20.1
marshmallow-sqlalchemy==1.1.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
python-jose==3.3.0
